+++ CHANGELOG.md
@@ -22,6 +22,8 @@
   [#331](https://github.com/cadeyrn/enterprise-policy-generator/issues/331)
 - Updated EnableTrackingProtection to support the SuspectedFingerprinting option, see
   [#327](https://github.com/cadeyrn/enterprise-policy-generator/issues/327)
+- Updated EnableTrackingProtection to support the BaselineExceptions and ConvenienceExceptions options, see
+  [#328](https://github.com/cadeyrn/enterprise-policy-generator/issues/328)
 - Updated Permissions policy to support the configuration of screen sharing requests, see
   [#329](https://github.com/cadeyrn/enterprise-policy-generator/issues/329)
 - Updated SanitizeOnShutdown policy: removed obsolete Downloads and OfflineApps options, added compatibility note for